Refactor recent performance changes in concurrency_oracle, #13
